optimizations

+++ python/test/test_backward_pass.py
@@ -86,11 +86,11 @@ def main():
     device_cpu = torch.device("cpu")
     device_gpu = torch.device("cuda") if has_cuda else None
 
-    Q_cpu = torch.rand(B, H, N, D_dim, device=device_cpu, dtype=torch.float32).contiguous()
-    K_cpu = torch.rand(B, H, N, D_dim, device=device_cpu, dtype=torch.float32).contiguous()
-    V_cpu = torch.rand(B, H, N, D_dim, device=device_cpu, dtype=torch.float32).contiguous()
+    Q_cpu = torch.rand(B, H, N, D_dim, device=device_cpu, dtype=torch.bfloat16).contiguous()
+    K_cpu = torch.rand(B, H, N, D_dim, device=device_cpu, dtype=torch.bfloat16).contiguous()
+    V_cpu = torch.rand(B, H, N, D_dim, device=device_cpu, dtype=torch.bfloat16).contiguous()
 
-    grad_out_cpu = torch.randn(B, H, N, D_dim, device=device_cpu, dtype=torch.float32).contiguous()
+    grad_out_cpu = torch.randn(B, H, N, D_dim, device=device_cpu, dtype=torch.bfloat16).contiguous()
 
     if has_cuda:
         Q_gpu = Q_cpu.to(device_gpu).contiguous()
